+++ src/apps/Art/hooks/index.ts
@@ -1,4 +1,5 @@
 import useBlockQueueProcessor from 'apps/Art/hooks/useBlockQueueProcessor';
+import useBroadcastBlock from 'apps/Art/hooks/useBroadcastBlock';
 import useCanvasArtworkAttributes from 'apps/Art/hooks/useCanvasArtworkAttributes';
 import useDetailsPageArtwork from 'apps/Art/hooks/useDetailsPageArtwork';
 import useIncomingTransferArtworks from 'apps/Art/hooks/useIncomingTransferArtworks';
@@ -10,6 +11,7 @@ import useValidArtworkAttributes from 'apps/Art/hooks/useValidArtworkAttributes'
 
 export {
   useBlockQueueProcessor,
+  useBroadcastBlock,
   useCanvasArtworkAttributes,
   useDetailsPageArtwork,
   useIncomingTransferArtworks,
